% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windowstack.R
\name{gridstack_widget_defaults}
\alias{gridstack_widget_defaults}
\title{See: https://github.com/gridstack/gridstack.js/blob/2ba4d412f348119e657a1d51a27fea360a2a9993/src/types.ts#L318}
\usage{
gridstack_widget_defaults()
}
\description{
See: https://github.com/gridstack/gridstack.js/blob/2ba4d412f348119e657a1d51a27fea360a2a9993/src/types.ts#L318
}
