% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window.R
\name{window_body}
\alias{window_body}
\alias{as.window_item}
\alias{is.window_item}
\title{window_body}
\usage{
window_body(
  ...,
  fillable = TRUE,
  fill = TRUE,
  class = NULL,
  style = NULL,
  id = NULL,
  wrapper = NULL
)

as.window_item(x)

is.window_item(x)
}
\arguments{
\item{x}{an object to test (or coerce to) a card item.}
}
\description{
\if{html}{\out{
<Add Description>
}}
}
\section{Functions}{
\itemize{
\item \code{as.window_item()}: Mark an object as a window item. This will prevent the
\code{\link[=window]{window()}} from putting the object inside a \code{wrapper} (i.e., a
\code{window_body()}).

}}
