% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window.R
\name{window_footer}
\alias{window_footer}
\title{window_footer}
\usage{
window_footer(
  ...,
  container = htmltools::div,
  class = NULL,
  style = NULL,
  id = NULL
)
}
\description{
\if{html}{\out{
<Add Description>
}}
}
