% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window.R
\name{window_toolbar}
\alias{window_toolbar}
\title{window_toolbar}
\usage{
window_toolbar(
  ...,
  container = htmltools::div,
  grabbable = TRUE,
  class = NULL,
  style = NULL,
  id = NULL
)
}
\description{
\if{html}{\out{
<Add Description>
}}
}
