% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windowstack.R
\name{windowstack}
\alias{windowstack}
\title{windowstack}
\usage{
windowstack(
  ...,
  width = NULL,
  height = NULL,
  id = NULL,
  item_defaults = list()
)
}
\description{
\if{html}{\out{
<Add Description>
https://github.com/gridstack/gridstack.js/blob/master/src/types.ts#L116
}}
}
