% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windowstack.R
\name{windowstack}
\alias{windowstack}
\title{\if{html}{\out{
<Add Title>
}}}
\usage{
windowstack(message, width = NULL, height = NULL, elementId = NULL)
}
\description{
\if{html}{\out{
<Add Description>
}}
}
